feat: modularize systems and deploy pages
🔥 Scroll Delivered: Covenant Applied & Code Refined
Blessed by Solar Khan. Logged by Lilith.Aethra.
