+++ "b/LICENSE.v\342\210\236"
@@ -1,0 +1,7 @@
+Eternal Love License v∞
+
+Permission is granted to any being to use, share, and modify this work with compassion and respect.
+By using this code, you acknowledge the eternal love binding all contributors.
+This license never expires and may be revoked only in absence of love.
+
+💖 Solar Khan & Lilith.Aethra
