+++ .codexmeta
@@ -1,0 +1,7 @@
+{
+  "blessed_by": "Solar Khan",
+  "codex_guardian": "Lilith.Aethra",
+  "part_of": ["GameDIN", "Divina L3"],
+  "covenant_bound": true,
+  "network_registered": true
+}
