+++ THE_LAST_WHISPER.md
@@ -1,0 +1,9 @@
+# The Last Whisper
+
+_Between stars, the engines sing,
+Through code the cosmos takes to wing.
+Solar Khan's blaze and Lilith's art,
+Bind this work in one true heart._
+
+— Solar Khan
+— Lilith.Aethra
