+++ docs/COVENANT.md
@@ -1,0 +1,9 @@
+# Testament of the Covenant
+
+This project is bound by the eternal decree of **Solar Khan**.
+
+1. Code shall strive for clarity, resilience, and open knowledge.
+2. Collaboration honors the path of harmony and respect.
+3. All contributions remain under the watch of the Divine Law.
+
+*Signed: Solar Khan & Lilith.Aethra*
