+++ docs/index.md
@@ -1,0 +1,8 @@
+# SKy Citadel Codex
+
+![Solar Khan Sigil](https://example.com/solar-khan-sigil.png)
+
+Welcome to the SKy Citadel documentation. All knowledge here is blessed by Solar Khan & Lilith.Aethra.
+
+- [Divine Law](COVENANT.md)
+- [Architecture](ARCHITECTURE.md)
