+++ gamedin/hub.json
@@ -1,0 +1,7 @@
+{
+  "name": "SKy Citadel",
+  "repository": "https://github.com/SolarKhan/SKyCitadel",
+  "covenant": true,
+  "divina_l3_pipeline": "v3",
+  "page": "https://SKyCitadel.SolarKhan.github.io"
+}
