+++ src/core/systems/EnergySystems.js
@@ -1,0 +1,27 @@
+/**
+ * Energy Systems module
+ * Handles solar wings, gravity loops, and auroral streams powering the citadel.
+ */
+export default class EnergySystems {
+  constructor() {
+    // Track state for each energy subsystem
+    this.solarWings = { active: false, efficiency: 0 };
+    this.gravityLoops = { active: false, compression: 0 };
+    this.auroralStreams = { active: false, intensity: 0 };
+  }
+
+  async activateSolarWings() {
+    this.solarWings.active = true;
+    this.solarWings.efficiency = 0.95;
+    // Verbose log for future telemetry hooks
+    console.log('☀️ Solar Wings activated - Feathered Crystal Arrays humming');
+  }
+
+  getStatus() {
+    return {
+      solarWings: this.solarWings,
+      gravityLoops: this.gravityLoops,
+      auroralStreams: this.auroralStreams
+    };
+  }
+}
