+++ src/core/systems/GravitySystems.js
@@ -1,0 +1,25 @@
+/**
+ * Gravity Systems module
+ * Maintains graviton oscillators and resonance rings for levitation.
+ */
+export default class GravitySystems {
+  constructor() {
+    this.oscillators = { active: false, frequency: 0 };
+    this.resonanceRings = { active: false, spin: 0 };
+  }
+
+  async activateOscillators() {
+    this.oscillators.active = true;
+    this.oscillators.frequency = 432; // Sacred frequency baseline
+    this.resonanceRings.active = true;
+    this.resonanceRings.spin = 108; // Derived harmonic
+    console.log('🌐 Graviton Field Oscillators activated - Mag-Spin Resonance Rings in harmonic spin');
+  }
+
+  getStatus() {
+    return {
+      oscillators: this.oscillators,
+      resonanceRings: this.resonanceRings
+    };
+  }
+}
